+++ CHANGELOG.md
@@ -33,6 +33,7 @@ Compatibility:
 * Fix `Encoding::Converter#primitive_convert` and raise `FrozenError` when a destination buffer argument is frozen (@andrykonchin).
 * Add `Module#undefined_instance_methods` (#3039, @itarato).
 * Add `Thread.each_caller_location` (#3039, @itarato).
+* Add `timeout` argument to `Thread::SizedQueue#push` (#3039, @itarato).
 
 Performance: